Replace validation import node method from controller into service.
